+++ python/GafferDispatch/__init__.py
@@ -44,6 +44,5 @@ from .Wedge import Wedge
 from .TaskContextVariables import TaskContextVariables
 from .TaskSwitch import TaskSwitch
 from .PythonCommand import PythonCommand
-from .FrameMask import FrameMask
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferDispatch" )